refactor: Refactor: split orbit-exec macos_sandbox.rs into focused subm… [ORB-00117]
Refactor: split orbit-exec macos_sandbox.rs into focused submodules

Planned-By: claude
